requirements.txt
google-cloud-storage==2.4.0
google-auth==2.6.6
python-dotenv==0.20.0
